Remove unnecessary script parsers
